Feature/chat

+++ src/main/java/com/green/greenGotell/domain/repository/QuestionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.green.greenGotell.domain.entity.QuestionEntity;
 
-public interface QuestionRepository extends JpaRepository<QuestionEntity, Long>{
+public interface QuestionRepository extends JpaRepository<QuestionEntity, Integer>{
 	QuestionEntity findByCategory(String category);
 }
